refactor(kikan): lift AppError + PlatformState from services/api [#507]
Prerequisite for S4.1 (platform wiring move). AppError becomes
kikan-owned so handlers like diagnostics/demo/backup_status can relocate
without creating an I4-violating dep on services/api.

- New kikan::app_error with AppError + IntoResponse (no vertical variants
  lifted; all original variants are HTTP-generic).
- New kikan::platform_state with PlatformState + MdnsStatus; outer
  SharedState exposes AppState::platform_state() for downstream use.
- SetupMode relocated from kikan::tenancy to kikan_types::profile to
  break the kikan-types -> kikan cycle introduced by kikan depending on
  kikan_types::error. Re-exported from kikan for source compatibility.
- services/api::error is now a thin re-export; services/api::discovery
  re-exports MdnsStatus from kikan.
- AppState.active_profile switched to Arc<RwLock<_>> so PlatformState is
  cheaply Clone-able per Axum FromRef requirements; all .read()/.write()
  call sites unchanged (auto-deref).
