+++ crates/kikan-types/Cargo.toml
@@ -9,7 +9,6 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 ts-rs = { workspace = true }
 mokumo-core = { workspace = true }
-kikan = { workspace = true }
 
 [dev-dependencies]
 chrono = { workspace = true }
