+++ crates/kikan-types/src/setup.rs
@@ -6,7 +6,7 @@ use ts_rs::TS;
 pub struct SetupStatusResponse {
     pub setup_complete: bool,
     #[ts(type = "\"demo\" | \"production\" | null")]
-    pub setup_mode: Option<kikan::SetupMode>,
+    pub setup_mode: Option<crate::SetupMode>,
     pub is_first_launch: bool,
     pub production_setup_complete: bool,
     pub shop_name: Option<String>,
@@ -24,14 +24,14 @@ pub struct DemoResetResponse {
 #[ts(export)]
 pub struct ProfileSwitchRequest {
     #[ts(type = "\"demo\" | \"production\"")]
-    pub profile: kikan::SetupMode,
+    pub profile: crate::SetupMode,
 }
 
 #[derive(Debug, Clone, Serialize, TS)]
 #[ts(export)]
 pub struct ProfileSwitchResponse {
     #[ts(type = "\"demo\" | \"production\"")]
-    pub profile: kikan::SetupMode,
+    pub profile: crate::SetupMode,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
