+++ crates/kikan/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod activity;
+pub mod app_error;
 pub mod app_handle;
 pub mod auth;
 pub mod backup;
@@ -9,11 +10,13 @@ pub mod error;
 pub mod graft;
 pub mod middleware;
 pub mod migrations;
+pub mod platform_state;
 pub mod profile_db;
 pub mod rate_limit;
 pub mod tenancy;
 
 pub use activity::{ActivityLogEntry, ActivityWriter, SqliteActivityWriter};
+pub use app_error::AppError;
 pub use app_handle::AppHandleShim;
 pub use boot::{BootConfig, DeploymentMode};
 pub use engine::{Engine, EngineContext, Sessions};
@@ -22,5 +25,6 @@ pub use error::{
 };
 pub use graft::{Graft, SelfGraft, SubGraft};
 pub use migrations::{GraftId, Migration, MigrationRef, MigrationTarget};
+pub use platform_state::{MdnsStatus, PlatformState, SharedMdnsStatus};
 pub use profile_db::{ActiveProfile, ProfileDb};
 pub use tenancy::{ProfileId, SetupMode, Tenancy};
