+++ crates/kikan/src/tenancy/profile_id.rs
@@ -1,46 +1,9 @@
-use serde::{Deserialize, Serialize};
 use std::fmt;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum SetupMode {
-    Demo,
-    Production,
-}
-
-impl SetupMode {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Demo => "demo",
-            Self::Production => "production",
-        }
-    }
-
-    pub fn as_dir_name(&self) -> &'static str {
-        self.as_str()
-    }
-}
-
-impl fmt::Display for SetupMode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Demo => write!(f, "demo"),
-            Self::Production => write!(f, "production"),
-        }
-    }
-}
-
-impl std::str::FromStr for SetupMode {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.trim().to_lowercase().as_str() {
-            "demo" => Ok(Self::Demo),
-            "production" => Ok(Self::Production),
-            other => Err(format!("unknown setup mode: {other}")),
-        }
-    }
-}
+// SetupMode lives in `kikan-types` so `kikan-types` does not depend on
+// `kikan` — necessary for `kikan` to depend on `kikan-types` (for the
+// `AppError`/`ErrorCode` types lifted in S4.0) without a dependency cycle.
+pub use kikan_types::SetupMode;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ProfileId(SetupMode);
